requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
openpyxl==3.1.0
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
Werkzeug==2.2.2
XlsxWriter==3.0.8
zipp==3.12.0
